requirements.txt
# PainPoint.er Search - Requirements

# Core dependencies
pandas>=1.3.0
tqdm>=4.62.0
python-dotenv>=0.19.0
selenium>=4.30.0
praw>=7.8.1

# AI API clients
google-generativeai>=0.3.0
openai>=1.0.0
anthropic>=0.8.0
azure-ai-inference>=1.0.0b9
azure-core>=1.32.0

# Analysis tools
textblob>=0.15.3
pandas>=2.2.3

# Optional: For development
pylint>=2.8.0
black>=21.5b2

requests